# user-service/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Service API",
    "version": "1.0.0",
    "description": "Servicio de gestión de usuarios protegido por JWT y CSRF. Autenticación delegada en Auth Service."
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Listar todos los usuarios (protegido por JWT)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de usuarios" },
          "401": { "description": "Token no proporcionado" },
          "403": { "description": "Token inválido" }
        }
      },
      "post": {
        "summary": "Crear nuevo usuario (protegido por CSRF)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "usuario@correo.com" },
                  "password": { "type": "string", "example": "secreto123" },
                  "role": { "type": "string", "example": "USER" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario creado correctamente" },
          "400": { "description": "Error de validación" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Obtener usuario por ID (protegido por JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Datos del usuario" },
          "404": { "description": "Usuario no encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar usuario por ID (protegido por JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Usuario actualizado" }
        }
      },
      "delete": {
        "summary": "Eliminar usuario por ID (protegido por JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Usuario eliminado" }
        }
      }
    },
    "/users/credentials": {
      "post": {
        "summary": "Validar credenciales (correo y password) - CSRF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "usuario@correo.com" },
                  "password": { "type": "string", "example": "secreto123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credenciales válidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "string" },
                    "email": { "type": "string" },
                    "role": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciales inválidas" },
          "404": { "description": "Usuario no encontrado" }
        }
      }
    },
    "/csrf-token": {
      "get": {
        "summary": "Obtener token CSRF",
        "responses": {
          "200": {
            "description": "Token CSRF emitido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csrfToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}